backend/src/test/java/org/cleviox/backend/service/EthereumServiceImplTest.java
package org.cleviox.backend.service;

import org.cleviox.backend.model.Balance;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;
import org.web3j.protocol.Web3j;
import org.web3j.protocol.core.Request;
import org.web3j.protocol.core.Response;
import org.web3j.protocol.core.methods.response.EthGetBalance;
import org.web3j.protocol.core.methods.response.EthSendTransaction;

import java.io.IOException;
import java.math.BigDecimal;
import java.math.BigInteger;

import static org.junit.jupiter.api.Assertions.*;
import static org.mockito.ArgumentMatchers.*;
import static org.mockito.Mockito.mock;
import static org.mockito.Mockito.when;

@ExtendWith(MockitoExtension.class)
class EthereumServiceImplTest {
/*
    @Mock
    private Web3j web3j;

    @InjectMocks
    EthereumServiceImpl ethereumService;

    @Test
    void testGetBalance_ValidAddress() {
        String validAddress = "0x123456789abcdef";
        BigInteger balanceWei = new BigInteger("1000000000000000000"); // 1 ETH in Wei
        Balance ethBalance = new Balance(new BigDecimal(1));

        Request<?, EthGetBalance> request = mock(Request.class);
        EthGetBalance ethGetBalance = mock(EthGetBalance.class);

        when(web3j.ethGetBalance(validAddress, any())).thenReturn(mock(Response.class));
        when(request.send()).thenReturn(ethGetBalance);

        Balance result = ethereumService.getBalance(validAddress);

        assertEquals(new BigDecimal("1.0"), result.balance());
    }

    @Test
    void testGetBalance_InvalidAddress() {
        String invalidAddress = "invalid_address";

        assertThrows(IllegalArgumentException.class, () -> ethereumService.getBalance(invalidAddress));
    }

    @Test
    void testGetBalance_IOException() throws Exception {
        // Arrange
        String validAddress = "0xf7E8E5E928bc7B3B2a1500a99a4D6Cc87307fc9a";

        // Mock the Request and EthGetBalance
        Request<?, EthGetBalance> ethGetBalanceRequest = mock(Request.class);
        EthGetBalance ethGetBalance = mock(EthGetBalance.class);

        // Use ArgumentMatchers.any() for handling type matching in generics
        when(web3j.ethGetBalance(any(String.class), any(DefaultBlockParameterName.class)))
                .thenReturn(ethGetBalanceRequest);

        when(ethGetBalanceRequest.send()).thenThrow(IOException.class);

        assertThrows(BalanceRetrievalException.class, () -> ethereumService.getBalance(validAddress));
    }




    @Test
    void testSendTransaction_ValidTransaction() throws Exception {
        String privateKey = "valid_private_key";
        String recipientAddress = "0xRecipientAddress123";
        BigDecimal amountInEth = new BigDecimal("0.1");

        EthSendTransaction ethSendTransaction = mock(EthSendTransaction.class);
        when(ethSendTransaction.getTransactionHash()).thenReturn("tx_hash_12345");
        when(web3j.ethSendRawTransaction(anyString())).thenReturn(ethSendTransaction);

        String transactionHash = ethereumService.sendTransaction(privateKey, amountInEth, recipientAddress);

        assertEquals("tx_hash_12345", transactionHash);
    }

    @Test
    void testSendTransaction_InsufficientFunds() {
        String privateKey = "valid_private_key";
        String recipientAddress = "0xRecipientAddress123";
        BigDecimal amountInEth = new BigDecimal("10"); // Assuming this is more than available funds

        // Setup the mock to throw the specific InsufficientFundsException or check a low balance
        when(ethereumService.getBalance(anyString())).thenReturn(new Balance(new BigDecimal("0.5")));

        assertThrows(InsufficientFundsException.class, () ->
                ethereumService.sendTransaction(privateKey, amountInEth, recipientAddress)
        );
    }
*/}